Fixing up inconsistencies in file extensions, rebase errors and documentation
