+++ drake_ros_core/README.md
@@ -6,17 +6,17 @@ It is similar to this ROS1 prototype [`gizatt/drake_ros_systems`](https://github
 
 ## Building
 
-This package has been built and tested on Ubuntu Focal with ROS2 Rolling, using a Drake nightly from April 2021.
+This package has been built and tested on Ubuntu Focal with ROS2 Rolling, using a Drake nightly or any stable releases after 14 Jan 2022.
 It may work on other versions of ROS and Drake, but it hasn't been tested.
 
 To build it:
 
 1. [Install ROS Rolling](https://index.ros.org/doc/ros2/Installation/Rolling/)
 1. Source your ROS installation `. /opt/ros/rolling/setup.bash`
-1. [Download April-ish 2021 Drake binary](https://drake.mit.edu/from_binary.html)
-1. Extract the Drake binary installation, install it's prerequisites, and [use this Python virutalenv trick](https://drake.mit.edu/python_bindings.html#inside-virtualenv).
-1. Activate the drake virtual environment
-1. Build it using Colcon, or using CMake directly
+1. [Download Drake binary](https://drake.mit.edu/from_binary.html), nightly or any stable releases after 14 Jan 2022.
+1. Extract the Drake binary installation, install it's prerequisites, and [use this Python virutalenv trick](https://drake.mit.edu/from_binary.html).
+1. Activate the drake virtual environment.
+1. Build it using Colcon, or using CMake directly.
     
     **Colcon**
     1. Make a workspace `mkdir -p ./ws/src`
