+++ drake_ros_core/package.xml
@@ -10,7 +10,6 @@
 
   <buildtool_depend>ament_cmake_ros</buildtool_depend>
 
-  <depend>pybind11-dev</depend>
   <depend>rclcpp</depend>
   <depend>rclpy</depend>
   <depend>rosidl_runtime_c</depend>
