+++ drake_ros_viz/.clang-format
@@ -38,7 +38,7 @@ IncludeCategories:
   - Regex:    '^<'
     Priority: 40
   # Your project's h files.
-  - Regex:    '^"drake_ros_core'
+  - Regex:    '^"drake_ros_viz'
     Priority: 50
   # Other libraries' h files (with quotes).
   - Regex:    '^"'
