+++ drake_ros_viz/CPPLINT.cfg
@@ -41,6 +41,12 @@ filter=+build/pragma_once
 # Disable cpplint's include order.  We have our own via //tools:drakelint.
 filter=-build/include_order
 
+# TODO(hidmic): uncomment when ament_cpplint updates its vendored cpplint
+# version to support the following filters. Also remove corresponding NOLINT
+# codetags in sources.
+## Allow private, sibling include files.
+#filter=-build/include_subdir
+
 # We do not care about the whitespace details of a TODO comment.  It is not
 # relevant for easy grepping, and the GSG does not specify any particular
 # whitespace style.  (We *do* care what the "TODO(username)" itself looks like
