+++ drake_ros_viz/include/drake_ros_viz/utilities/name_conventions.h
@@ -11,8 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#ifndef DRAKE_ROS_TF2__UTILITIES__NAME_CONVENTIONS_HPP_
-#define DRAKE_ROS_TF2__UTILITIES__NAME_CONVENTIONS_HPP_
+#pragma once
 
 #include <string>
 #include <unordered_set>
@@ -20,22 +19,24 @@
 #include <drake/geometry/scene_graph_inspector.h>
 #include <drake/multibody/plant/multibody_plant.h>
 
-namespace drake_ros_tf2 {
+namespace drake_ros_viz {
 namespace utilities {
 
-std::string GetTfFrameName(
-    const drake::geometry::SceneGraphInspector<double>& inspector,
-    const std::unordered_set<const drake::multibody::MultibodyPlant<double>*>&
-        plants,
-    const drake::geometry::FrameId& frame_id);
-
-std::string GetTfFrameName(
+/** Retrieve conventional marker namespace for a given scene geometry.
+ @param[in] inspector inspector for a given SceneGraph's data.
+ @param[in] plants a set of MultibodyPlant instances from which to derive
+   semantically meaningful marker namespaces, if possible.
+ @param[in] geometry_id target geometry ID.
+ @returns the namespace for the marker
+ @pre `geomtry_id` is associated with a registered geometry in the SceneGraph
+   implied by `inspector`.
+ @pre all `plants` are associated with the SceneGraph implied by `inspector`.
+ */
+std::string GetMarkerNamespace(
     const drake::geometry::SceneGraphInspector<double>& inspector,
     const std::unordered_set<const drake::multibody::MultibodyPlant<double>*>&
         plants,
     const drake::geometry::GeometryId& geometry_id);
 
 }  // namespace utilities
-}  // namespace drake_ros_tf2
-
-#endif  // DRAKE_ROS_TF2__UTILITIES__NAME_CONVENTIONS_HPP_
+}  // namespace drake_ros_viz
